Store api token in keychain
It wasn't safe to store the pinboard API token using UserDefaults as these are stored as plain text.
Instead we store the api token in the keychain and tried to set it up so the same keychain item can
be shared between the main app and the Safari app extension without requiring the user's password.

It is not clear yet if sharing the keychain item without requiring password works. We see both apps in the keychain
access control list but we are still asked for a password. This might be because our app is not
signed yet.