Change way to get selected text in injected script
We use a 'selectionchange' event listener and store the selected text, on every change, in a variable
instead of getting the selected text on demand when opening the popover window.
We do this to work around the deselect issue when the toolbar button is pressed.

We store the selected text in the 'selectedText' variable but only empty it if we got
two or more consecutive empty selections. This to work around the deselect issue when the toolbar button is pressed.
Following sequences work should work:

a. 1. User selects text.  2. selection is lost when user presses toolbar button (outside of our control)
b. 1. User deselects text. 2. selection is lost when user presses toolbar button (outside of our control)

In a. the user selected text will appear in the Description field as expected as we only got a single empty selection (2.)
In b. the Description field will be empty as expected because we got 2 consecutive empty selections (1. and 2.)