Support editing of already bookmarked URL
Fixes #18

When opening the popup we will check if the page link was already submitted earlier.
If this is the case we indicate when it was added initially and populate the previously
submitted properties.

The user can update any properties and add it again which will update the existing pinboard link.