Populate URL and title immediately when popover appears
Since introducing support for retrieving and updating previously
saved bookmarks there was a delay when populating the URL and title because we
were waiting for a response of the pinboard api to find out if the URL already
was bookmarked previously.

The default case should be that the URL doesn't exist
yet and so populating the URL and description right away makes sense and minimizes the lag.
The consequence is that it these properties might still get overwritten in case it was
already bookmarked before.