Tag suggestions implementation
First implementation of tag suggestions. Fixes #6.
It shows a dropdownlist with suggestions based on the user's previously used tags.
The suggesions are ordered based on occurence. The most used tags are suggested first.
You can select a tag using the arrow down key and select it using the arrow right key.
Tab completion unfortunately doesn't work.

The implementation has also other limitations.
It only suggests a dropdown list with suggestions for the first 2 typed characters for every tag.
This means that you won't get suggestions anymore if you type more than 2 characters for a tag.
The reason for this limitation is that the trigger for showing the suggestions is any text change and
selecting a tag from the dropdown list triggers a text change which on its turn updates the suggestions to only the
selected item which is not useful. The restriction of only showing suggestions for the first
two characters works around this behaviour as the majority of the selected tags will have > 2 characters.